FAL-1760 avoid instantiating FileSpecs multiple times
This is only the first step of FAL-1760.

The question remains why sometimes NodeGroup is used and in other places
properties are passed through, instead of using getNodeGroup().
Do properties change later on?
Should the file specs and the file handler be built only once (for
example in setNodeGroup) ?

This at least seems to reduce the constructor calls from 22 to 4 in the
ctool end to end test yaml parsing.